Make signals_to_port_fields patch work with newer PORTs
The signals_to_port_fields was originally written with only the
"old"-style PORT register blocks in mind (which only have PORTn, PINn,
and DDRn registers).

Newer AVR devices instead have a more complex register block for PORTs
which contains not only registers like the old-style where each bit
corresponds to a pin, but also pin-specific registers.  Update the patch
to only touch the former type of register and leave the latter as-is
without erroring out.