Add patch to remove prefixes from register names
In some ATDF descriptions, register names have a peripheral specific
prefix.  For example on ADC there is a list of registers named
`ADC_<name>`.

Add an optional post-processing patch which drops such a common prefix
for all registers of a peripheral.  It can be enabled by passing

	--auto-patches remove_register_common_prefix

to atdf2svd.  The patch will search for the longest common prefix of
each peripherals' registers.  The prefix must end in an `_` to be
accepted.