Ensure we don't generate empty enumerated values
Make sure we never generate `ValueRestriction::Enumerated` for an empty
list of enumerated values.  This will become important when switching to
svd-rs as it will refuse accepting such value restrictions.

So far, this situation was only observed for fields where we filter out
invalid enumerated values.  But the check will also cover hypothetical
ATDF files which have an empty list of enumerated values for some field.

Fixes: 62e5b6051951 ("Drop invalid enumerated values with a warning")