Emit `alternateGroup` element in SVD for registers with multiple modes
This patch adds the capability to emit an `alternateGroup` element in
the resulting SVD when a register of a peripheral has multiple modes.
This is handy for the modern timer peripherals in the atxmegas and new
attinys where a timer can be either set to 16 bit single mode or you can
split it into two independent 8 bit timers.

The resulting registers are prefixed with the mode already by `svd2rust`
as of version v0.18.0. For example in an attiny817 for peripheral TCA0
there will be `single_ctrl{a,b,c}` and `split_ctrl{a,b,c}` registers.