Introduce a default-enabled `cli` feature
With our biggest user, `avr-device` switching to using this crate as a
library rather than a CLI tool, it's time to somewhat pivot our
dependency management towards this usecase.  There is no need to pull in
all the CLI frontend dependencies for library users.

Introduce a new `cli` feature for the CLI interface.  This feature is
enabled by default so the tool can still be built without issue as
needed.  Library-only users like `avr-device` can disable the `cli`
feature to reduce the number of unnecessary dependencies.